Reorganize ServerError class to a new folder
