+++ src/server/errors/ServerError/ServerError.ts
