+++ src/server/errors/ServerError/types.ts
